fix removing funs file
